backend/requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.9
aiosignal==1.3.1
alembic==1.13.3
altair==4.2.2
aniso8601==9.0.1
annotated-types==0.7.0
anyio==4.6.0
appdirs==1.4.4
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.8.1
asttokens==2.0.5
async-lru==2.0.4
async-timeout==4.0.3
attrs==23.2.0
Babel==2.11.0
backoff==2.2.1
bcrypt==4.2.0
beautifulsoup4==4.12.3
bert-score==0.3.13
bleach==4.1.0
blinker==1.8.2
blis==1.0.1
boto3==1.35.34
botocore==1.35.34
Brotli==1.0.9
build==1.2.2
cachetools==5.5.0
catalogue==2.0.10
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
chroma-hnswlib==0.7.6
click==8.1.7
cloudpathlib==0.20.0
cloudpickle==2.2.1
cohere==5.11.1
colorama==0.4.6
coloredlogs==15.0.1
colorlog==6.8.2
comm==0.2.1
confection==0.1.5
contourpy==1.3.0
cycler==0.12.1
cymem==2.0.8
databricks-sdk==0.33.0
dataclasses-json==0.6.7
datasets==2.2.1
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.14
dill==0.3.9
distro==1.9.0
dnspython==2.7.0
docker==7.1.0
durationpy==0.9
en_core_web_sm==3.8.0
entrypoints==0.4
et-xmlfile==1.1.0
eval_type_backport==0.2.0
exceptiongroup==1.2.2
executing==0.8.3
Faker==30.8.0
fastapi==0.115.0
fastavro==1.9.7
fastjsonschema==2.16.2
favicon==0.7.0
filelock==3.16.1
Flask==3.0.3
flatbuffers==24.3.25
fonttools==4.54.1
fqdn==1.5.1
frozenlist==1.4.1
fsspec==2024.6.1
gitdb==4.0.11
GitPython==3.1.43
google-ai-generativelanguage==0.6.6
google-api-core==2.20.0
google-api-python-client==2.147.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-generativeai==0.7.2
google-pasta==0.2.0
googleapis-common-protos==1.65.0
graphene==3.3
graphql-core==3.2.4
graphql-relay==3.2.0
greenlet==3.1.1
grpcio==1.66.2
grpcio-status==1.62.3
gunicorn==23.0.0
h11==0.14.0
htbuilder==0.6.2
httpcore==1.0.6
httplib2==0.22.0
httptools==0.6.1
httpx==0.27.2
httpx-sse==0.4.0
huggingface-hub==0.25.1
humanfriendly==10.0
idna==3.10
importlib-metadata==6.11.0
importlib_resources==6.4.5
ipykernel==6.28.0
ipython==8.27.0
ipywidgets==8.1.2
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
jiter==0.5.0
jmespath==1.0.1
joblib==1.4.2
json5==0.9.6
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter==1.0.0
jupyter_client==8.6.0
jupyter-console==6.6.3
jupyter_core==5.7.2
jupyter-events==0.10.0
jupyter-lsp==2.2.0
jupyter_server==2.14.1
jupyter_server_terminals==0.4.4
jupyterlab==4.2.5
jupyterlab-pygments==0.1.2
jupyterlab_server==2.27.3
jupyterlab-widgets==3.0.10
kiwisolver==1.4.7
kubernetes==31.0.0
langchain==0.2.16
langchain-community==0.2.17
langchain-core==0.2.41
langchain-openai==0.1.20
langchain-text-splitters==0.2.4
langcodes==3.4.1
langsmith==0.1.131
language_data==1.2.0
lxml==5.3.0
Mako==1.3.5
marisa-trie==1.2.1
Markdown==3.7
markdown-it-py==3.0.0
markdownlit==0.0.7
MarkupSafe==2.1.5
marshmallow==3.22.0
matplotlib==3.9.2
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==2.0.4
mlflow==2.16.2
mlflow-skinny==2.16.2
mmh3==5.0.1
mock==4.0.3
monotonic==1.6
more-itertools==10.5.0
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.17
munch==2.5.0
murmurhash==1.0.10
mypy-extensions==1.0.0
nbclient==0.8.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.1
nltk==3.9.1
notebook==7.2.2
notebook_shim==0.2.3
numpy==1.26.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
oauthlib==3.2.2
onnxruntime==1.19.2
openai==1.51.0
openpyxl==3.1.5
opentelemetry-api==1.27.0
opentelemetry-exporter-otlp-proto-common==1.27.0
opentelemetry-exporter-otlp-proto-grpc==1.27.0
opentelemetry-instrumentation==0.48b0
opentelemetry-instrumentation-asgi==0.48b0
opentelemetry-instrumentation-fastapi==0.48b0
opentelemetry-proto==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-semantic-conventions==0.48b0
opentelemetry-util-http==0.48b0
orjson==3.10.7
overrides==7.4.0
packaging==23.2
pandas==2.2.3
pandocfilters==1.5.0
parameterized==0.9.0
parso==0.8.3
pathos==0.3.3
pexpect==4.8.0
pillow==10.4.0
pip==24.2
platformdirs==4.3.6
plotly==5.24.1
ply==3.11
posthog==3.7.0
pox==0.3.5
ppft==1.7.6.9
preshed==3.0.9
prometheus-client==0.14.1
prompt-toolkit==3.0.43
proto-plus==1.24.0
protobuf==4.25.5
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==17.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.21
pydantic==2.9.2
pydantic_core==2.23.4
pydantic-settings==2.5.2
pydeck==0.9.1
Pygments==2.18.0
pymdown-extensions==10.11.2
pymongo==4.10.1
pyparsing==3.1.4
PyPika==0.48.9
pyproject_hooks==1.2.0
PyQt5==5.15.10
PyQt5-Qt5==5.15.15
PyQt5_sip==12.15.0
pysbd==0.3.4
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
python-json-logger==2.0.7
pytz==2024.2
PyYAML==6.0.2
pyzmq==25.1.2
qtconsole==5.6.0
QtPy==2.4.1
ragas==0.1.21
referencing==0.35.1
regex==2024.9.11
requests==2.32.3
requests-mock==1.12.1
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
responses==0.18.0
retrying==1.3.4
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.2
rouge==1.0.1
rpds-py==0.20.0
rsa==4.9
s3transfer==0.10.2
safetensors==0.4.5
sagemaker==2.232.2
sagemaker-core==1.0.10
sagemaker-mlflow==0.1.0
schema==0.7.7
scikit-learn==1.5.2
scipy==1.14.1
semantic-router==0.0.70
Send2Trash==1.8.2
sentence-transformers==3.2.1
setuptools==75.1.0
shellingham==1.5.4
sip==6.7.12
six==1.16.0
smart-open==7.0.5
smdebug-rulesconfig==1.0.1
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.5
spacy==3.8.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.35
sqlparse==0.5.1
srsly==2.4.8
st-annotated-text==4.0.1
st-theme==1.2.3
stack-data==0.2.0
starlette==0.38.6
streamlit==1.39.0
streamlit-aggrid==1.0.5
streamlit-camera-input-live==0.2.0
streamlit-card==1.0.2
streamlit-embedcode==0.1.2
streamlit-extras==0.5.0
streamlit-faker==0.0.3
streamlit-image-coordinates==0.1.9
streamlit-keyup==0.2.4
streamlit-pills==0.3.0
streamlit-toggle-switch==1.0.2
streamlit-vertical-slider==2.5.5
sympy==1.13.1
tabulate==0.9.0
tblib==3.0.0
tenacity==8.5.0
terminado==0.17.1
thinc==8.3.2
threadpoolctl==3.5.0
tiktoken==0.8.0
tinycss2==1.2.1
together==1.3.3
tokenizers==0.20.0
toml==0.10.2
tomli==2.0.1
toolz==1.0.0
torch==2.5.0
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
transformers==4.45.2
triton==3.1.0
trulens==1.1.0
trulens-core==1.1.0
trulens-dashboard==1.1.0
trulens_eval==1.1.0
trulens-feedback==1.1.0
trulens-providers-openai==1.1.0
typer==0.12.5
types-python-dateutil==2.9.0.20241003
types-requests==2.32.0.20240914
typing_extensions==4.12.2
typing-inspect==0.9.0
tzdata==2024.2
uri-template==1.3.0
uritemplate==4.1.1
urllib3==2.2.3
uvicorn==0.31.0
uvloop==0.20.0
validators==0.34.0
wasabi==1.1.3
watchdog==5.0.3
watchfiles==0.24.0
wcwidth==0.2.5
weasel==0.4.1
webcolors==24.8.0
webencodings==0.5.1
websocket-client==1.8.0
websockets==13.1
Werkzeug==3.0.4
wheel==0.44.0
widgetsnbextension==4.0.10
wrapt==1.16.0
xxhash==3.5.0
yarl==1.13.1
zipp==3.20.2
